Add files via upload
Black and white bmp of the college logo

Follow the below steps to import BMP file in Fusion 360 PCB workspace:
Open the Fusion 360 PCB workspace.
Go to the 2D layout file and select the Automation tab.
Choose Run ULP option, browse and select from the list the ULP file called "import-bmp" in Fusion 360.
Select the .bmp file from a computer location to import the file to Fusion 360.
Scan used colors or select them manually and click OK.
Adjust settings of the imported image (on which layer number it will be putted e.g. 1 - it will be on the top copper, change units to e.g. Mill or change scale ).
Click OK.

https://knowledge.autodesk.com/support/fusion-360/troubleshooting/caas/sfdcarticles/sfdcarticles/How-to-import-BMP-file-into-Fusion-360-PCB-workspace.html